Create README.md
Researchers have recently proposed that ``moralistic'' religions---those with moral doctrines, moralistic supernatural punishment, and lower emphasis on ritual---emerged as an effect of greater individual-level wealth and material security. This proposal predicts that individuals with ``slow life history'' strategies will be more attracted to moralistic traditions as a means to judge those with ``fast life history'' strategies. We investigate this hypothesis with a data set consisting of 592 individuals from eight diverse societies. Our sample represents a wide range of traditions, including world religions such as Buddhism, Hinduism and Christianity, but also local traditions such as beliefs in animism, ancestor worship, and worship of spirits associated with nature. We first test for evidence of an association between material security, years of formal education, and number of children. We then test whether or not these demographic factors predict the moral concern, punitiveness, attributed knowledge-breadth of, and frequency of ritual devotions towards two deities in each society. While we find support for the predictions that material security and education are associated with fewer children, we find no reliable evidence of a relationship between number of children, material security, or formal education and the individual-level beliefs and behaviors typifying moralistic religions.